your_teeth/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
tensorflow-cpu==2.15.0
numpy==1.24.3
pillow==10.1.0
matplotlib==3.8.2
opencv-python-headless==4.8.1.78
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1